use token to protect routes for deleting messages/notifications and submitting/deciding deck submissions

+++ server/routes/message.js
@@ -117,60 +117,87 @@ messageRouter.patch('/:messageId/add-to-read', getUserIdFromJWTToken, async (req
             const updatedMessage = await Message.findByIdAndUpdate(req.message._id, {$addToSet: {read: req.body.readerId}}, {new: true});
             res.status(200).send({messageId: updatedMessage._id, read: updatedMessage.read});
         } else {
-            res.status(404).send("User is not authorized to update the read array of this message");
+            res.status(403).send("Only recipients of a message may mark it as read");
         }
     } catch (err) {
         res.status(500).send(err.message);
     }
 });
 
-messageRouter.patch('/:messageId', async (req, res, next) => {
+messageRouter.patch('/:messageId',  getUserIdFromJWTToken, async (req, res, next) => {
     try {
+        const foundUser = await User.findById(req.userId);//only get the fields needed by the cases
+        console.log({userIdInMessageRouter: req.userId});
+        if(!foundUser.messages.received.includes(req.message._id)) {
+            res.status(403).send("Unauthorized");
+            return;
+        }
         switch(req.body.messageType) {
-            case "DeckSubmission": //should this be DeckSubmission
+            case "DeckSubmission":
                 //first make sure hasn't already been approved/denied by another admin
                 const foundDeckSubmissionMessage = await DeckSubmission.findById(req.message._id).populate("targetDeck", "name");
+                
+                // then make sure the deciding user is an admin of the group (may have initially been when message sent but then removed)
+                const foundGroup = await Group.findById(foundDeckSubmissionMessage.targetGroup, "administrators");
+                if(!foundGroup.administrators.includes(req.userId)) {
+                    res.status(403).send("Only group administrators may approve or deny submitted decks");
+                    return;
+                }
+                
 
-                //if it has, send back the acceptanceStatus and (if approved) the added deck id
+                //if it has already been approved/denied, send back the acceptanceStatus and (if approved) the added deck id
                 if(foundDeckSubmissionMessage.acceptanceStatus !== "pending") {
-                    res.status(200).send({
-                        acceptanceStatus: foundDeckSubmissionMessage.acceptanceStatus,
-                        deckName: foundDeckSubmissionMessage.targetDeck.name,
-                        //only send back the _id if approved because otherwise the deck would have been deleted and the id would no longer exist
-                        ...(foundDeckSubmissionMessage.acceptanceStatus === "approved" && {deckId: foundDeckSubmissionMessage.targetDeck._id})
-                    });
+                    res.status(409).send(`This deck has already been ${foundDeckSubmissionMessage.acceptanceStatus}`);
                 } else {
                     //otherwise first update the acceptanceStatus of the DeckSubmission message
                     await DeckSubmission.findByIdAndUpdate(req.message._id, {acceptanceStatus: req.body.decision});
 
                     // get the deck name to send back for the message in case deck/id get deleted if submission denied
-                    const submittedDeck = await Deck.findById(req.body.deckId, "name creator");
+                    const submittedDeck = await Deck.findById(foundDeckSubmissionMessage.targetDeck, "name creator");
 
                     //handle add/delete deck based on approval/denial decision
                     if(req.body.decision === "approved") {
-                        const updatedDeck = await Deck.findByIdAndUpdate(req.body.deckId, {approvedByGroupAdmins: true}, {new: true});
-                        const updatedGroup = await Group.findByIdAndUpdate(req.body.groupId, {$push: {decks: updatedDeck._id}});
-                        //possibly exclude the approver's id
-                        const otherGroupMembers = updatedGroup.members.filter(memberId => memberId.toString() !== req.body.decidingUserId && memberId.toString() !== req.message.sendingUser.toString());
-                        await User.updateMany({_id: {$in: otherGroupMembers}}, {$push: {notifications: await DeckAddedNotification.create({targetDeck: updatedDeck, targetGroup: foundDeckSubmissionMessage.targetGroup, read: false})}});
+                        const updatedDeck = await Deck.findByIdAndUpdate(foundDeckSubmissionMessage.targetDeck, {approvedByGroupAdmins: true}, {new: true});
+                        const updatedGroup = await Group.findByIdAndUpdate(foundDeckSubmissionMessage.targetGroup, {$push: {decks: updatedDeck._id}});
+                        //possibly exclude the approver's id;
+            
+                        const otherGroupMembers = updatedGroup.members.filter(memberId => memberId.toString() !== req.userId && memberId !== req.message.sendingUser);
+                        
+                        const bulkOperations = await Promise.all(otherGroupMembers.map(async (memberId) => {
+                            const notification = await DeckAddedNotification.create({
+                                targetDeck: updatedDeck, 
+                                targetGroup: foundDeckSubmissionMessage.targetGroup, 
+                                read: false
+                            });
+
+                            return {
+                                updateOne: {
+                                    filter: {_id: memberId},
+                                    update: {$push: {notifications: notification}}
+                                }
+                            };
+                        }));
+                        
+                        await User.bulkWrite(bulkOperations);
+
                     } else if(req.body.decision === "denied") {
-                        await Deck.findByIdAndDelete(req.body.deckId)
+                        await Deck.findByIdAndDelete(foundDeckSubmissionMessage.targetDeck)
                     }
 
                     const deckDecisionMessage = new DeckDecision({
-                        sendingUser: req.body.decidingUserId,
+                        sendingUser: req.userId,
                         receivingUsers: [submittedDeck.creator. _id],
                         acceptanceStatus: req.body.decision,
                         comment: req.body.comment,
                         deckName: submittedDeck.name,
-                        targetDeck: req.body.deckId,
-                        targetGroup: req.body.groupId,
+                        targetDeck: foundDeckSubmissionMessage.targetDeck,
+                        targetGroup: foundDeckSubmissionMessage.targetGroup,
                         targetUser: submittedDeck.creator
                     });
 
                     const savedDeckDecisionMessage = await deckDecisionMessage.save();
                     await User.findByIdAndUpdate(foundDeckSubmissionMessage.sendingUser, {$push: {"messages.received": savedDeckDecisionMessage}});
-                    await User.findByIdAndUpdate(req.body.decidingUserId, {$push: {"messages.sent": savedDeckDecisionMessage}});
+                    await User.findByIdAndUpdate(req.userId, {$push: {"messages.sent": savedDeckDecisionMessage}});
 
                     res.status(200).send({sentMessage: {_id: savedDeckDecisionMessage._id, read: savedDeckDecisionMessage.read, messageType: savedDeckDecisionMessage.messageType}, acceptanceStatus: req.body.decision});
                 }
@@ -212,7 +239,7 @@ messageRouter.patch('/:messageId', async (req, res, next) => {
                             //         throw err;
                             //     }
                             //   });
-                              
+                            
                             // await User.bulkWrite(bulkOps);
                         }  
 
@@ -241,41 +268,44 @@ messageRouter.patch('/:messageId', async (req, res, next) => {
             default:
                 break;
         }    
-
     } catch (err) {
         console.error(err);
         res.status(500).send(err.message);
     }
 });
 
-messageRouter.delete("/:messageId", async (req, res, next) => {//test a message that had multiple receivingUsers (delete one from sender first then from one Receiver then sender then rest of receivers)
-    console.log({query: req.query});
-    const updateObj = {
-        ...(req.query.direction === "received" && {$pull: {receivingUsers: req.query.deletingUserId}}),
-        ...(req.query.direction === "sent" && {$set: {sendingUserDeleted: true}})
-    }
-    console.log({updateObj});
-
+messageRouter.delete("/:messageId", getUserIdFromJWTToken, async (req, res, next) => {
     try {
-        const updatedMessage = await Message.findByIdAndUpdate(req.message._id, updateObj, {new: true});
-        console.log({updatedMessage});
-        
-        let updatedUser;
-        if(req.query.direction === "received") {
-            updatedUser = await User.findByIdAndUpdate(req.query.deletingUserId, {$pull: {"messages.received": req.message._id}}, {new: true});    
-        } else if(req.query.direction === "sent") {
-            updatedUser = await User.findByIdAndUpdate(req.query.deletingUserId, {$pull: {"messages.sent": req.message._id}}, {new: true});
-        } else {
-            throw new Error("invalid direction submitted");
-        }
-        console.log({updatedUser});
+        const deletingUser = await User.findById(req.userId, "_id messages");
 
-        //if sending user and all receivingUsers have deleted the message from their inbox, delete from database
-        if(!updatedMessage.receivingUsers.length && updatedMessage.sendingUserDeleted) {
-            await Message.findByIdAndDelete(req.message._id);
-            res.status(200).send("Successfully deleted");
+        if(deletingUser?.messages[req.query.direction].includes(req.message._id)) {
+            const updateObj = {
+                ...(req.query.direction === "received" && {$pull: {receivingUsers: deletingUser._id}}),
+                ...(req.query.direction === "sent" && {$set: {sendingUserDeleted: true}})
+            }
+            const updatedMessage = await Message.findByIdAndUpdate(req.message._id, updateObj, {new: true});
+            console.log({updatedMessage});
+            
+            let updatedUser;
+            if(req.query.direction === "received") {
+                updatedUser = await User.findByIdAndUpdate(deletingUser._id, {$pull: {"messages.received": req.message._id}}, {new: true});    
+            } else if(req.query.direction === "sent") {
+                updatedUser = await User.findByIdAndUpdate(deletingUser._id, {$pull: {"messages.sent": req.message._id}}, {new: true});
+            } else {
+                throw new Error("invalid direction submitted");
+            }
+        
+    
+            //if sending user and all receivingUsers have deleted the message from their inbox, delete from database
+            if(!updatedMessage.receivingUsers.length && updatedMessage.sendingUserDeleted) {
+                await Message.findByIdAndDelete(req.message._id);
+                res.status(200).send("Successfully deleted");
+            } else {
+                res.status(200).send("message removed from user's messages");
+            }
+
         } else {
-            res.status(200).send("message removed from user's messages");
+            res.status(403).send("Only recipients of a message may deleted it");
         }
         
     } catch (err) {
